refactor: Instead of rendering the Phaser canvas into hardcoded DOM elements, it now renders in a React-rendered component by using componentDidMount (useEffect hook). Also moved game config to its own file in our config folder
