+++ client/components/index.js
@@ -14,3 +14,4 @@ export {default as RollingDice} from './rollingDice'
 export {default as PanelL} from './panelL'
 export {default as PanelR} from './panelR'
 export {default as GameViewTitle} from './gameViewTitle'
+export {default as GameView} from './gameView'
