+++ client/routes.js
@@ -9,7 +9,7 @@ import {
   HomePage,
   ChooseCharacter,
   RollingDice,
-  GameViewTitle
+  GameView
 } from './components'
 import {me} from './store'
 
@@ -29,7 +29,7 @@ class Routes extends Component {
         {/* Routes placed here are available to all visitors */}
         <Route exact path="/" component={HomePage} />
         <Route exact path="/create" component={ChooseCharacter} />
-        <Route exact path="/game" component={GameViewTitle} />
+        <Route exact path="/game" component={GameView} />
         <Route exact path="/dice" component={RollingDice} />
         <Route path="/login" component={Login} />
         <Route path="/signup" component={Signup} />
