+++ public/index.html
@@ -9,11 +9,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <div id="gameView">
-      <div id="panelL"></div>
-      <div id="theGame"></div>
-      <div id="panelR"></div> 
-    </div>
   </body>
 </html>
